WIP - [Feature] Add var to disable respawn instances
This is a Draft/WIP - I am not as familiar with perl, but rather than maintaining a table of zones that should not have access to respawning instances, I added a variable to pass through the OfferStandardInstance.

This should exclude the respawning option when set to "True" and do a hard return with a message if someone tries to bypass the hail.

This has NOT been tested and should be reviewed by @Catapultam prior to being accepted as he may have other ideas for the design.